Fix auth_state for Zero-To-Jupyterhub on Globus Auth
Zero-To-Jupyterhub differs slightly in that it packages auth_state
data in JSON before sending it off to a user's newly created pod.
This will fail if any of the data isn't JSON-serializable. This
change simply encodes it to be JSON friendly. The regular Jupyterhub
does not send auth_state data to the spawner as JSON so it isn't
affected by this bug.